[fix] windows blocks on resize and sends multiple resize events
This fix will make it so that the event loop keeps track of the
resize width and height as the window resizes. Then when it gets
the window message that it has finished resizing it will send
the resize event once.

In the future this may be fixed for a continuous resize event as the
user resizes, but that will require manually dispatching the resize
event.